MRJ: Applied:
In the KDE-HTMLwidget the tokenizer segfaults, when HTMLTokenizer::write is
called after HTMLTokenizer::end and before HTMLTokenizer::begin has allocated a
new buffer. This is the reason of some crashes in kfm, where a io-job seems to
continue to write data, when no buffer is allocated. The following patch fixes
the segfault, eventually a warning-message should be printed,  because it seems
to be a bug in kfm.

Lars Heete
--------------------------------------------------------------
--- htmltoken.cpp.orig  Tue Jun 16 18:33:01 1998
+++ htmltoken.cpp       Thu Jun 18 18:32:26 1998
@@ -125,11 +125,11 @@
     // this function.
     char *srcPtr = 0L;

     KCharsets *charsets=KApplication::getKApplication()->getCharsets();

-    if ( str == 0L )
+    if ( str == 0L || buffer == 0L)
        return;

    const char *src = str;

    // If we have <pre> and get a '\t' we need to know